requirements.txt
flask
google-genai
python-dotenv
youtube_transcript_api
